move DOM functions to util/
